no_zk:
- Removed P from proof creation function; it was initially added because we needed to generate alpha, and so we needed to seed the PRNG with some data.
This however can be done outside of the function, before the prover calls create function.

Why?

If we keep P inside of the function, all higher level protocols will need to pass P into the create function. With this, they only need to pass the transcript which contains P from some source.

- Made verify() as method on NoZK

Why?

The NoZK struct will be embedded in the AlmZK struct. Having verify as a method on NoZk makes the API nicer when we want to verify a IPA_ALM_ZK. We would use a period '.' to call verify instead of '::'

- Made verify() return a boolean

Why?

The NoZK struct will be embedded in the AlmZK struct. We want a way to propagate the results from NoZK to AlmZK; if NoZK fails verification, then AlmZK also fails.